Refactor PageItem
subtypes of pageitem define the uuid/permalink for each kind
of page item.

e.g. an organization has relations such as current_team and news.
current_team is a paged list of Person nodes on CrunchBase,
identified via their permalink which lives in the 'path' returned in the
json.
assume that we want more information about the first perosn in the
current team:
    person = organization.current_team.get(1)  # will get you a PersonPageItem
    permalink = person.permalink  # will get you the permalink of person
    crunchbase_api.person(permalink)  # will get you the details